chore: proper material-ui theme config
