+++ src/theme/overrides.js
@@ -1,0 +1,17 @@
+import { DEFAULT_PALETTE } from './palettes'
+
+const overrides = {
+  MuiSnackbar: {
+    root: {
+      maxWidth: '515px',
+      width: '100%'
+    }
+  },
+  MuiMobileStepper: {
+    root: {
+      background: DEFAULT_PALETTE.background.main
+    }
+  }
+}
+
+export { overrides }
