+++ src/theme/palettes.js
@@ -1,0 +1,21 @@
+const DEFAULT_PALETTE = {
+  type: 'dark',
+  primary: {
+    main: '#cfd8dc',
+    light: '#ffffff',
+    dark: '#9ea7aa',
+    contrastText: '#fff'
+  },
+  secondary: {
+    main: '#ffe082',
+    light: '#ffffb3',
+    dark: '#caae53',
+    contrastText: '#fff'
+  },
+  background: {
+    main: '#181a1b',
+    light: '#1f2123'
+  }
+}
+
+export { DEFAULT_PALETTE }
